Remove

+++ scripts/reset.sh
@@ -7,9 +7,12 @@ echo Removing migrations
 rm -rf drizzle
 echo "Dropping db"
 
-dropdb -f --if-exists kidarr
-echo "Creating db"
-createdb kidarr
+if $NODE_NEV ne "production"; then
+  echo "Dropping production db"
+  dropdb -f --if-exists kidarr
+  echo "Creating production db"
+  createdb kidarr
+fi
 
 bun db:generate
 bun db:migrate
